r8499: This patch contains changes to the rdma_bw.c to allow use of RDMA CM. This has been successfully tested with Ammasso iWARP cards and IBM eHCA IB cards.
Added an option (-c|--cma) to enable use of RDMA CM.

Added a new structure (struct pp_data) containing the user parameters as well
as other data required by most of the routines. This makes it convenient to
pass the parameters between various routines.

All outputs to stdout/stderr are prefixed with the process-id. This helps to
sort the output when multiple servers/clients are run from the same machine.

Signed-off-by: Pradipta Kumar Banerjee <bpradip@in.ibm.com>
Signed-off-by: Michael S. Tsirkin <mst@mellanox.co.il>