write_bw : Enabled cycle buffer writing to remotre address.
This feature significantly improves the BW in Nahaelem system as it allows the
Remote machine to write the data to the memory simultaneously.
The fix is enable only if message size is smaller than 4K , because bigger sized messages
are allready ay line rate.
Basiclly the change is to read/write the data from different address every time, and allign
it to 64B address. because of that we can now exploid the "CACHE_LINE" mechanism in Nahaelem.

Example of the improvement on 1KB message size :
Before ~ 820 MB/s ,  After ~ 2930 MB/s.

Signed-off-by: Ido Shamay <idos@dev.mellanox.co.il>