perftest_resources : "ctx_set_mtu" is added to the function library.
If the user request for specific mtu , the method checks if the size is legal
(MTU sizes are 256, 512, 1024, 2048, 4096) and if it's not above the active_mtu.
Otherwise , it sets the mtu to be the active_mtu configured by the driver.
In UD connection type , the method checks if the message size is not above the mtu.

Signed-off-by: Ido Shamay <idos@dev.mellanox.co.il>`