perftest : Aligned all read / write from memory buffer to 64B addresses.
We now set the scatter gathered elements to read / write allways from addresses that
are aligned with 64B , in order to enable the cpu to only cache line opperations.

This feature include UD packets. when a QP receive a UD packet , we make sure , the data
will be written to 64B aligned address , taking account the 40B GRH section , which is
written (or skiped) in the first 40B of the message.

signed-off-by: Ido Shamay <idos@dev.mellanox.co.il>