perftest: Added support of rdma_cm to exchange test data.
Now you can use, in all tests, the "-R" flag to instruct the test to pass all of the relevant data it usually
passed through TCP/IP with the rdma_cm module. It allows the test to run using only the IB/RoCE interface,
without the need of TPC/IP interface as well.

signed-off-by: Ido Shamay <idos@dev.mellanox.co.il>