perftest_parameters: Changed Inline and Max Inline computation and method.
From now, user can choose any Max Inline he chooses, up to maximum in driver (and not 400).
Latency and ConnectIB default are now optimal values calculated by statistics.
For BW tests default is still 0. Unknown devices (which are not known to support this feature) have allways MAX inline of 0.
If you want to add a device to supported devices list, please send a me a patch or the device ID and I'll add it.

signed-off-by: Ido Shamay <idos@dev.mellanox.co.il>