raw_ethernet_bw: Added forwarding test option (-v flag) to raw_etherent_bw.
This great feature acts as and Ethernet forwarding application, it reflects back all the messages it receive.
The payload is not changed and only source MAC is replaced with dest MAC in given packet and sent back to origin.
For small, it uses all of perftest performance mechanisms of handling multiple packets simultaneously.

Signed-off-by: Yuval Atias <yuvala@mellanox.com>
Code review: Ido Shamay <idos@mellanox.com>