rdma_cm feature: Bug fix for UD transport type in rdma_cm (-R) mode.
The UD feature with rdma_cm based QPs was broken since d66d7888ec4513ca0fd826216c64e20129d0bef8 (moving to perftest-2).
We tracked the bug with a binary search and found that when moving from set_send_wqes (static in send_lat.c) to ctx_set_send_wqes, we
forgot to handle the UD in rdma_cm mode (which needed a different configuration in QPN and QKEY).

signed-off-by: Ido Shamay <idos@dev.mellanox.co.il>