all latnecy tests: Added Duration support.
Now all latency tests have the duration feature that was introduced earlier in bw tests.
It runs the test for a fixed duration of seconds, displaying the average latency throughout this period.
With this feature the tests does not support for median,min,max results and also
histogram and unsorted results disply features (-H and -C flags in latency tests) are obselete.
Usage is the same as in BW tests: -D <duration in sec> [-f <margin of sample>].

Signed-off-by: Ido Shamay <idos@dev.mellanox.co.il>