New parameters for L2/3/4 in raw_ethernet tests
Added local and remote paramter for mac,ip nad port.
Client defines local as the source and remote as dest parameters for packet creation.
Server defines local as the dest and remote as source for incomig packets.
all local new parameters values should refer to the self params and remote for the paired.
avoid usage of old and new parameters format.
add support in case of --raw_ipv6

Signed-off-by: Zohar Ben Aharon <zoharb@mellanox.com>