Avoid configuring the MRs with 1's
Avoid setting the value '1' in the MR. If this happens in a write latency
test then the server will send two consecutive packets, regardless of the
client's state. This can cause the application to hang - If the client
reaches the busy-wait loop after the second write then it'll keep waiting
for the value of the first write forever.
to avoid that we set 0's for the MR only for WRITE LAT tests.

Signed-off-by: Ram Amrani <Ram.Amrani@cavium.com>
Signed-off-by: Zohar Ben Aharon <zoharb@mellanox.com>