Fix HW rate limit for RoCE devices
RoCE currently doesn't support HW rate limit. This was no problem for all rates
since the query function returned error values (2), instead for 2.5Gb/s which is
defined to be "2". Since this is the same as the error value, no problem was
detected and perftest tries to run with HW rate limit instead of existing or
falling back to SW rate limit.

Fixed by checking the device's link type. If it is RoCE we behave as if we got an
error or rate mismatch, i.e. fall back to SW rate limit or exit if HW was explicitly
provided.

Signed-off-by: Tal Gilboa <talgi@mellanox.com>