configure: fix ibv_wr_api enable argument
The syntax for AC_ARG_ENABLE is:
  AC_ARG_ENABLE(option-name, help-string, action-if-present, action-if-not-present)

The mistake here is to consider the two actions as action-if-enabled and
action-if-disabled.

Scenario:
  ./configure --enable-ibv_wr_api
would end up running action-if-present and cause the new API to be
disabled.

Fix:
Since the default for autoconf when no action is defined is to set a
special variable named with the enable_ prefix, remove the custom
actions and let autoconf handle it.

Fixes: 705ed2e ("Made ibv_wr_api default during the build")
Signed-off-by: Firas Jahjah <firasj@amazon.com>