Fix iWARP connection failures - Modify backlog passed to rdma_listen.
The size of the backlog affects the time it takes to run rdma_listen in iWARP.
The iWARP stack allocates work entries based on the number.
When passing 0, the default is used which is 1K, this translates to the iWARP
rdma_listen taking ~40 MS leading to connection establishment timeouts
(when the client attempts to connect before the server is ready ).
There is no need to send the maximum backlog as the perftest is written in a
one pending connection establishment anyway, therefore, modify the rdma_listen
to use "1" as the backlog.

Signed-off-by: Michal Kalderon <michal.kalderon@marvell.com>