Perftest: Add support for encryption/decryption in perftest
Add a feature that encrypts/decrypts the data before
sending it. The feature is supported only on crypto devices.

For using this feature you should first follow the instruction
in README file.

Adding two files that is mandatory to use them if willing to
use this feature

The following flags were added:
    1) --aes_xts - Runs traffic with AES_XTS feature (encryption/decryption).
    2) --encrypt_on_tx Runs traffic with encryption on tx & decryption
        on rx (opposite is the default).
    3) --sig_before Puts signature on data before encrypting it (default
        after).
    4) --aes_block_size=<512,520,4048,4096,4160> (default 512) - every
        aes_block_size of the message we sent is encrypted with
        different tweak.
    5) --data_enc_keys_number=<number of data encryption keys> (default 1).
    6) --kek_path path to the key encryption key file.
    7) --credentials_path path to the credentials file.
    8) --data_enc_key_app_path path to the data encryption key app.

This support was added to the following applications:
ib_read_bw, ib_write_bw, ib_send_bw, ib_read_latency and ib_send_latency.

Signed-off-by: Hassan Khadour <hkhadour@nvidia.com>