Perftest: Fix possible performance issue caused by conditional if's
Fix performance concern  in case the wrap of AES_XTS code is defined
but the traffic sent is not encrypted.

This possible issue is solved in compilation time by adding a
function parameter "enc" that indicates if the data sent is
encrypted/decrypted to the inline _new_post_send function
declaration and by using a pointer to this function.

This fix will make every case of sending data use a different
flow of post send wr optimized to it - reducing the run rime.

Signed-off-by: Hassan Khadour <hkhadour@nvidia.com>