Perftest: Fix huge pages overflow
Huge pages with size > 2^31  allocation fails due to
that the variables containing the message size and
huge page size which is an alignment of the message size
are int typed which caused an overflow.

This commit fixes this issue by changing the type
of the variable mentioned above to uint64_t to
prevent the overflow.

Signed-off-by: Hassan Khadour <hkhadour@nvidia.com>