configure: Cleanup CUDA configury
The current approach to enabling CUDA support in the bandwidth
benchmarks with an environment variable that takes that path to an
include file is extremely fragile.  As a low-hanging fruit, this commit
brings the CUDA configuration closer to ROCm, but relying on a flag to
enable the feature and AC_SEARCH_LIBS to update lib flags, or on
--with-cuda allowing a user to provide a custom installation path.
Ideally, we should rely on dlopen()'ing the symbols we need from these
libraries so a single build of perftest can work on systems with and
without CUDA SDK, but that's to come later.

Signed-off-by: Raghu Raja <raghu@enfabrica.net>