Revert "Perftest:support get_cycles for riscv"
This reverts commit 1780311f7bb12dee2b53dddbc8076d4ca4ea65bc.

This commit doesn't solve the problem of get_cycles().

At first, it uses CSR_TIMEx registers to acquire CPU cycles but
from the RISC-V specification (riscv-spec-v2.2 and higher) this register is
responsible for real-time-clock. On SiFive HiFive Unmatched board it
counts just microseconds from an arbitrary start time.

Due to the RISC-V ISA specification (riscv-privileged-v1.9 and higher)
the second problem is that if we want to use RISC-V counter registers in U-mode
we have to enable that counter in MUCOUNTEREN register.

This path was checked on the SiFive HiFive Unmatched board and showed that tests
doesn't work correctly.

Signed-off-by: v.v.mitrofanov <v.v.mitrofanov@yadro.com>