Perftest: fix broken IPv4 checksum due to type aliasing
The IP checksum implementation in raw_ethernet_lat aliases a uint16_t with
the structured IPv4 header. This generates incorrect code under strict
aliasing rules when certain optimizations are enabled and ip_checksum() is
inlined into gen_ipv4_header(). The consequence of this is that the IPv4
checksum gets incorrectly computed.

Using a union to perform the type conversion is a safe way to avoid
running afoul of strict aliasing in gcc [1].

[1] https://gcc.gnu.org/onlinedocs/gcc-12.1.0/gcc/Optimize-Options.html#Type-punning

Signed-off-by: Edwin Peer <edwin@enfabrica.net>