Perftest: calculate IPv4 checksum 32-bits at a time
Optimize the IPv4 checksum calculation in ip_checksum() by summing
32-bit integers instead of 16-bits at a time. Also drop the carry
check inside the loop, it was superflous even for the 16-bit case,
since the ten 16-bit adds over the length of the header could never
have accumulated that many carries. For the 32-bit case, grow the
size of the accumulating variable to 64-bits. Any carries into the
upper 32-bits will also be folded in at the end.

Signed-off-by: Edwin Peer <edwin@enfabrica.net>