Perftest: fix ending conditions for non-standard -t/-q/-n values
It was revealed that in certain combination of tx_depth, connection
number, cq_mod and iteration values the code wrongly determines
ending conditions.

Especially, in the while loop condition a wrong comparison was used
in which a mix of signed and unsigned variables are summed or
subtracked without minding the resulting sign:

(ctx->scnt[index] - ctx->ccnt[index] + user_param->post_list) <=
    (user_param->tx_depth)
In this case when scnt is lower than ccnt by more than post_list
we end up with a high number compared to tx_depth just because
the scnt is unsigned.

In another case, for multiple qp involved, when cq_mod is not a divider
of iters we may end up in situation when:
  totccnt >= tot_iters
and there is an i-th connection with ctx->ccnt[$i] < iters
In such scenario the test will not be able to complete.

Commands for reproduction:
ib_read_bw -R -s 1024 -n 10000 -q 5 -t 12
ib_send_bw -a -b -R -t 83 -r 41
OS: FreeBSD 12/13

Sponsored-by: Intel Corporation
Signed-off-by: Bartosz Sobczak <bartosz.sobczak@intel.com>