Perftest: close completion channel bug fix
On DC connection, perftest creates 2 CQ's on the client side, a CQ that
is associated with the QP receive queue and one with the send queue.
Currently, perftest code is trying to close the completion channel that
is associate with both of the CQ's. The channel closing failed because
the receive CQ is still open.
This commit fixes the issue by destroying the receive CQ before closing
the channel.

Signed-off-by: Shmuel Shaul <sshaul@nvidia.com>