Perftest: Fix ib_send_bw with cuda+rdma_cm+bidir+ud
When using ib_send_bw with RDMA CM + UD connection + bidirectional +
use_cuda it cause to segmentation fault.
the fault caused by attempt to accese a buffer
that allcoated on the GPU memory inside the
'rdma_cm_initialize_ud_connection_parameters()' function.
The commit fix the error by copy the buffer to the host
memory and then use the copy.

Fixes: None
Signed-off-by: Shmuel Shaul <sshaul@nvidia.com>