Perftest: Fix raw_ethernet_bw with GPUDirect
When using raw_ethernet_bw with use_cuda flag,
perftest code try to access buffer that allocated
on the GPU buffer.
this happen when it try to generate an ethernet header
and when printing the info of the header.
This commit fix these issues by copy the buffer contnet
to memory that allocate on the CPU.

Signed-off-by: Shmuel Shaul <sshaul@nvidia.com>