After the establishment of the connection, disconnect_left is incremented on both the server and client sides.
The rdma_disconnect API is called on both the server and client sides
during connection termination based on the disconnect_left value,
but the disconnect_left is missing in the server, which results in
the server not calling disconnect during termination, which causes an error.

1. either the client-side disconnect fails.

2. Despite the fact that the rdma_disconnect was successful,
   the RDMA_CM_EVENT_DISCONNECTED was not sending the target to the client,
   which is only sent if the rdma_disconnect is called at the server side.
   As a result, the count of disconnect_left does not decrease,
   and the client continues to wait for RDMA_CM_EVENT_DISCONNECTED indefinitely.

By adding the disconnect_left server-side parameter, both the server and client will
call the disconnect API, and the client will send RDMA_CM_EVENT_DISCONNECTED to the
server and the server will send RDMA_CM_EVENT_DISCONNECTED to the client,
thus decrementing disconnect_left on both sides.