Use separate completion channels for send and receive
Using a single channel for both send and receive completions leads to
cases where the test performs a TX operation, waits for it to finish
but gets woken up due to an RX event (and vice versa).
This patch splits the completion channel to two, for the send and
receive CQs.

One exception is the bidirectional test, which actually expects either a
send or receive completion in order to advance. This is handled by
calling select() on both fds.

Signed-off-by: Gal Pressman <galpress@amazon.com>
Signed-off-by: Michael Margolin <mrgolin@amazon.com>