Fix race in non-rdmacm ctx_close_connection()
When rdmacm is not used, ctx_close_connection() does a handshake and
then sends "done" over the socket before closing it. This is racy and
can lead to a spurious error:

    HOST A                                  HOST B

    ctx_close_connection()
                                            ctx_close_connection()
      ctx_hand_shake() [ succeeds ]
                                              ctx_hand_shake() [ succeeds ]
                                              write(sockfd, "done")
                                              close(sockfd)

      write(sockfd, "done") <-- fails since HOST B has closed the
                                socket and replies with a TCP RST

Fix this simply by deleting the write(). The ctx_hand_shake() already
ensures the two sides are in sync and can proceed to the close().

Signed-off-by: Roland Dreier <roland@enfabrica.net>