Perftest: fix completion count with special combination of tx_depth/cq_mod/iterations bug
When cq_mod/tx_depth is not a divider of iterations
perftest run may end up in situation when:
total completion count >= total iterations
and there is an i-th connection with ctx->completion count[$i] < iterations.
In such scenario the test will not be able to complete.
This commit handle this bug by adding the needed checks.

Signed-off-by: Bartosz Sobczak <bartosz.sobczak@intel.com>
Signed-off-by: Shmuel Shaul <sshaul@nvidia.com>