Display a warning if BW peak measurement was disabled implicitly
In BW measurements which run, e.g., with a specified duration, BW peak
measurement is disabled implicitly, and a peak BW 0 is displayed in the
results' summary. This can be confusing to the user, who does not choose
to run without peak-bw explicitly, and makes the user assume that
something went wrong during the run. To mitigate this, a warning message
is displayed now.

Signed-off-by: Leonid Bloch <lb.workbox@gmail.com>