Perftest: Fix failure in creating cq when create cq ex is not supported by providers
The previous TD lock-free patch introduced ibv_create_cq_ex() to
perftest, but it breaks the provider devices which do not support
this API yet. For these devices calling this API leads to an errno
of EOPNOTSUPP. So add a check of errno, and if it is EOPNOTSUPP,
use ibv_create_cq() as a fallback.

Fixes: b6f957f6bc6c ("Perftest: Add support for TD lock-free mode")
Tested-By: Selvin Xavier <selvin.xavier@broadcom.com>
Signed-off-by: Junxian Huang <huangjunxian6@hisilicon.com>