Added correct list of required packages(Depends) for perftest.deb
Perftest requires the rdma-core and libibumad packages for proper operation, but they are missing from the dependency list.
As a result, perftest can be installed successfully, and its tools can be used, but some will not work without the libibumad package.
This patch adds the missing dependencies to the perftest package.