Perftest: Fix perform warm up process stuck
In perform_warm_up mode, if the length of post_list is 1 and the
message size is less than or equal to 8192, all send_flags in WRs
are 0 and CQEs will not be generated since IBV_SEND_SIGNALED is
not set. As a result, the perform_warm_up process will stuck in
an infinite poll-CQ loop.

Set IBV_SEND_SIGNALED in this case to requiring CQE, and clear the
flag after post_send_method to avoid affecting subsequent tests.

Fixes: 56d025e4f19a ("Allow overriding CQ moderation on post list mode (#58)")
Signed-off-by: Guofeng Yue <yueguofeng@h-partners.com>
Signed-off-by: Junxian Huang <huangjunxian6@hisilicon.com>