Use CQE QP number
Perftest assumes that all WQEs are created with the QP number as their
wr_id. This creates a situation where WQEs do not have unique IDs. As a
result, in various areas of the code, wr_id field from wc is used as the
index to the relevant QP. The usage of wr_id is replaced with the
actual QP index that also exists in the wc struct.

Reviewed-by: Daniel Kranzdorf <dkkranzd@amazon.com>
Reviewed-by: Yonatan Nachum <ynachum@amazon.com>
Signed-off-by: Nitzan Lavy <nlavy@amazon.com>