Perftest: Add option for send BW with immediate
Add an option to run send BW tests while using the immediate field.
The option can be turned on by using the flag '--send_with_imm".
Data will be written to imm_data field of the WQE on the sender side,
and it will be verified that it arrives correctly on the receiver side.

Reviewed-by: Nitzan Lavy <nlavy@amazon.com>
Reviewed-by: Michael Margolin <mrgolin@amazon.com>
Signed-off-by: Daniel Kranzdorf <dkkranzd@amazon.com>