Perftest: Write BW pattern data validation
Extending write_bw data validation scenario to offer pattern data
validation.

The data buffer of which the packets consists of contains serial data in
granularity of DWORD. The pattern is taken from an input file, using the
flag --payload_file_path.

The format of the pattern input file is DWORD in hex, followed by a
comma. An example: 0xaaaaaaaa,0xbbbbbbbb,...

The patterns will be used in cyclic way to fill the whole data buffer.

Reviewed-by: Firas Jahjah <firasj@amazon.com>
Reviewed-by: Daniel Kranzdorf <dkkranzd@amazon.com>
Signed-off-by: Nitzan Lavy <nlavy@amazon.com>