Perftest: Write BW serial data validation
Extending write_bw data validation scenario to offer serial data
validation.

The data buffer of which the packets consists of contains serial data in
granularity of DWORD.

The user can set the initial value from which the data will increment
from by using the flag --data_start_value. By default the data will have
a starting value of 0.

As an example, if the user used the flag --start_data_value 64, it will
result in the following data: 00000040 00000041 00000042 ...

Reviewed-by: Firas Jahjah <firasj@amazon.com>
Reviewed-by: Daniel Kranzdorf <dkkranzd@amazon.com>
Signed-off-by: Nitzan Lavy <nlavy@amazon.com>