Perftest: Write BW random data validation
Adding a capability in perftest to validate the integrity of the data
transferred. For now, this functionality works only for write_bw test
scenarios. Currently, the data generated is random. In the following
commits data validation is expanded to serial and pattern data types.

Activate random data validation scenario by using the following flag:
	--data_validation random

Data validation implementation has a few limitations. The following
flags and conditions must be applied when running data validation:
	post_list (-l) == tx_depth (-t)
	recv_post_list (--recv_post_list) == rx_depth (-r)
	--write_with_imm

Size enforcement of the QPs is required to overcome some perftest
limitations for reusing of data buffers of WQEs.

Turning on the immediate flag is also required as the immediate field is
used as a hint for the data.

Reviewed-by: Firas Jahjah <firasj@amazon.com>
Reviewed-by: Daniel Kranzdorf <dkkranzd@amazon.com>
Signed-off-by: Nitzan Lavy <nlavy@amazon.com>